+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Fixed
+- core: `AttributeError` with `asyncio.tasks._gather` on Python 3.6 (#23).
+
 ## [1.1.1] - 2024-11-05
 
 ### Added