+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [1.2.0] - 2024-11-26
 
 ### Added
 - settings: print list parameters as comma-separated list of values for more
@@ -97,7 +97,8 @@ and this project adheres to
 
 ## [1.0.0] - 2024-04-08
 
-[unreleased]: https://github.com/rackslab/RFL/compare/v1.1.1...HEAD
+[unreleased]: https://github.com/rackslab/RFL/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/rackslab/RFL/releases/tag/v1.2.0
 [1.1.1]: https://github.com/rackslab/RFL/releases/tag/v1.1.1
 [1.1.0]: https://github.com/rackslab/RFL/releases/tag/v1.1.0
 [1.0.3]: https://github.com/rackslab/RFL/releases/tag/v1.0.3